fix: add ISR support 60s revalidate
Fix issue #59
ISR for people, projects, and single project pages.

temp
