+++ src/utils/index.ts
@@ -1,0 +1,2 @@
+export { getImgUrlFromAttachmentObj } from './image';
+export { sortPeople } from './people';
